Moved a lot more of the param slider stuff to another file
